requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
gensim>=4.3.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
pytest>=7.3.0
pytest-django>=4.5.0
django-cors-headers>=4.0.0
coreapi>=2.3.3 